libfetchers/mercurial: default to current revision on local repos
When using the Mercurial fetcher on a local repository without explicitly
specifying a branch or revision, previously always the tip of the default
branch would be fetched. This is likely unexpected by the user, and
inconsistent with the Git fetcher as well as the dirty case. To reduce
surprises and restore consistency, fetch the currently checked out revision
instead.

Change-Id: Id6d58f958b710f7a9661dc66ad4ddcb8d06a0cdd