libstore: don't return promises from LocalDerivationGoal::startBuilder
this is misleading at best and dangerous at worst. callers expect all
exceptions thrown here to signal build startup failure, and errors in
the build itself to be signaled through the promise. since this isn't
a virtual method it will always return the same kind of promise, i.e.
we can just call handleChildOutput in the one caller of startBuilder.

Change-Id: I65d06f96ec05a3fcdf2050dbbd87c417c308ecc9