libstore: always load localstore public keys
it's dirt cheap to do on average, the complexity of sticking this under
a lock is not worth it. we may want to split this into not loading both
kinds of keys since private keys are rarely needed, but since only root
or the daemon are likely to have access to them it shouldn't be urgent.

Change-Id: I691aed100c9cc4ca32ab7e99a37b5be7d5c25e93