libstore: remove withFramedSink exception print
it's entirely not necessary. any time we'd print this exception we're
already being called as part of stack unwinding, and printing the way
it's done here causes crashes when the error is a user pressing ^C to
abort the current operation. printing all errors (including ones that
aren't interruptions) only causes unnecessary noise in terminal logs.

fixes fj#646

Change-Id: I965ff57bc9ecf7dff35963d7056aa3a5d8a83e39