libstore: simplify LocalStore::createTempRootsFile
we now create a temporary file, lock it, move it into place, and return.
if any of these steps (except creation) fails the GC must have found and
deleted our new file before we got the chance to use it. we don't need a
lock around file creation anymore, we especially don't need a mutex that
covers an flock call, and we don't need the GC to clobber what it finds.
(clobbering *is* kept for compatibility with older nix implementations.)

if we ever leave trash around it'll be cleaned up by some future gc run.

Change-Id: I078ae55fb7a9dfcdae28d6917d138842188ed528