subprojects/nix-eval-jobs: build tests in Hydra jobs
This exposes the `pytest` suite as a Nix build. Previously, this didn't
exist and upstream just runs this in GHA. As a result, the test was
actually broken since the assumptions about the directory structure have
changed.

To not require network access, the nixpkgs used by the Lix flake is
injected into the test. The evaluation will be done in a chroot store
since the `openStore` call would otherwise need write access to
/nix/store.

Change-Id: Ib4ff5f0319ce3aabb3b7243a076ea169e27e3533