libutil: mirror nar structure in nar visitor structure
this is an unfortunate but necessary intermediate state in which we can
mirror nar structure from the visitor side, but not yet from the parser
side. to move from paths to entry names we must first relocate names of
entries out of the entries themselves, and into their parent directory.

Change-Id: I2ea898230952a4f631eea67c66ac8dcbdfb19636