libutil: abstract nar parser for other kinds of sources
now we can feed the nar parser from something that isn't a Source, such
as a kj async stream. we will not be able to wrap the nar members quite
as nicely in that case since async generators are not a thing yet. once
we need that we can come up with some async generator mechanism though.

Change-Id: If9f5d7865c18e4293aa037b9acd29718e080b3e4