libstore: don't use parseDump to discard nars
this is a step towards using nar parser visitors only where we need nar
parser visitors. the performance impact of this should be insignificant
and perhaps slightly positive since we no longer allocate anything when
discarding a file member. mostly it is a step towards a NARParseVisitor
that has no default behavior; we have only three implementations total,
and most of them don ot even *use* most of the default implementations.

Change-Id: I3466d2a77500778eb152f56ebd441cf1be223dc6