libstore: extract source implementation from NarIndexer
this is conflating concerns (position tracking in the input stream and
actually unpacking its contents) in an unhelpful way and is preventing
us removing full-path access requirements from the visitor concept. we
want to move from reconstructing and then deconstructing paths to only
passing individual entry names to nar parsers, who can then build full
paths from the nar root only if the need those to function. not all of
our implementations need this; only one of the three does, one ignores
the paths entirely, and the third even takes the reconstruction apart.

Change-Id: If249f13c5b9ae5b9e4e4a3777875ef8364ee6229