libstore: split Store::computeStorePathForPath
the two variants compute their store path hashes very differently, and
the flat version ignores the filter passed in by the user entirely. in
a future change we want to move filters out of store methods entirely,
splitting them now will make that task a lot easier when we are ready.

Change-Id: I5f23a26ca08e81923f33adf687056d0d464c2cb8